feat: refactor pool
